Generate Confluence macro for an expand block

Generate Confluence macro for an expand block

confl_macro_expand(title, body)

Arguments

title
defines the text that appears next to the expand/collapse icon

body
this HTML content will be visible when someone clicks the macro title

Value

HTML as string

Note

content needs to be HTML, so look at commonmark::markdown_html, pander::pander and eg xtable for doing the conversion before passing to confluence_expand

References

https://confluence.atlassian.com/doc/expand-macro-223222352.html

Examples

if (FALSE) {
confl_macro_expand(
  'Example block',
  commonmark::markdown_html(pander::pander_return(list(a = list(b = 4), c = 2))))
}