Publish R Markdown Document to 'Confluence'

Knit and post a given R Markdown file to 'Confluence'.

confl_create_post_from_Rmd(Rmd_file, interactive = NULL, params = NULL, ...)

confluence_document(
  title = NULL,
  space_key = NULL,
  parent_id = NULL,
  type = c("page", "blogpost"),
  toc = FALSE,
  toc_depth = 7,
  supported_syntax_highlighting = getOption("conflr_supported_syntax_highlighting"),
  update = NULL,
  use_original_size = FALSE,
  interactive = NULL
)

Arguments

Rmd_file
Path to an .Rmd file.

interactive
If FALSE, shiny interface is not launched.

params
If provided, a list of named parameters that override custom params in the YAML front-matter.

...
Arguments passed to confluence_documents().

title
Title of the post.

space_key
The space key to find content under.

parent_id
The page ID of the parent pages.

type
The content type to return. Default value: page. Valid values: page, blogpost.

toc
If TRUE, include a table of contents in the output.

toc_depth
The max level of headers to include in the table of contents.

supported_syntax_highlighting
A named character vector of supported syntax highlighting other than default (e.g. c(r = "r")).

update
If TRUE, overwrite the existing page (if it exists).

use_original_size
If TRUE, use the original image sizes.

Details

title, type, space_key, parent_id, toc, toc_depth, update, and use_original_size can be specified as confluence_settings item in the front-matter of the Rmd file to knit. The arguments of confl_create_post_from_Rmd() overwrite these settings if provided.