REST Wrapper for the ContentService

REST Wrapper for the ContentService

confl_list_pages(
  type = c("page", "blogpost", "comment", "attachment"),
  limit = 10,
  start = 0,
  spaceKey = NULL,
  title = NULL,
  expand = NULL
)

confl_get_page(id, expand = "body.storage")

confl_post_page(
  type = c("page", "blogpost"),
  spaceKey,
  title,
  body,
  ancestors = NULL,
  image_size_default = 600
)

confl_update_page(id, title, body, image_size_default = 600)

confl_delete_page(id)

Arguments

type
The content type to return. Default value: page. Valid values: page, blogpost.

limit
The limit of the number of items to return, this may be restricted by fixed system limits.

start
The start point of the collection to return.

spaceKey
The space key to find content under.

title
The title of the page to find. Required for page type.

expand
A comma separated list of properties to expand. To refer the nested contents, use periods. (e.g. body.storage,history).

id
ID of the content.

body
The HTML source of the page.

ancestors
The page ID of the parent pages.

image_size_default
The default width of images in pixel. If NULL, images are displayed in their original sizes.

See also